requirements.txt
azure-storage-blob==12.14.1
pytest==7.3.1
pytest-cov==4.0.0
pytest-timeout==2.1.0
numpy==1.24.4
pandas==1.5.3
azure-identity==1.14.0
